dialyzer: Introduce option --no_opaque_types
The -Wno_opaque suppresses warnings, but the analyses are still done
with opaque types present.

The new option --no_opaque_types, on the other hand, handles the
-opaque() attribute exactly as the -type() attribute, effectively
turning all opaque type into non-opaque types.

This is useful for debugging, and also if Dialyzer's bugs make the
analyses meaningless or slow.